extracts deployment lofic into a helper module
